+++ src/App/Model/BuildingTask.hs
@@ -7,7 +7,7 @@ import App.Model.Installation (Installation)
 
 data BuildingTask = BuildingTask
   { installation :: Installation
-  , quantity :: Int
+  , quantity :: Double
   , finishTime :: Time Int
   }
   deriving (Show, Generic)
